pyproject.toml
[project]
name = "attention-approximation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "blobfile>=3.0.0",
    "datasets>=4.0.0",
    "google>=3.0.0",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "libcirkit>=0.2.1",
    "matplotlib>=3.10.5",
    "nbformat>=5.10.4",
    "safetensors>=0.6.2",
    "sentencepiece>=0.2.1",
    "tiktoken>=0.11.0",
    "torchvision>=0.23.0",
    "transformers>=4.56.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["attention_approximation"]

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "W", "N", "UP", "B", "C90", "RUF"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ["py310"]
